XORG_FONT_REQUIRED_PROG: should stop when program is missing #14436
The macro was designed to test for a program and, when missing, stop
the configuration. The first parameter to the macro is a variable
name which is used by AC_PATH_PROG to store the program path.
The test failed to reference the variable content, it tested the
variable name which is never blank.

BDFTOPCF MKFONTSCALE MKFONTDIR COMPRESS and UCS2ANY were affected.

Reviewed-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Tested-by: Jesse Adkins <jesserayadkins@gmail.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>